better style
